Update apt-notifier.py
Start a singleshot timer (50 msec.) after closing Synaptic or the 'View and Upgrade' window.
While the singleshot timer is running, don't process clicks to start Synaptic, or to open the 'View and Upgrade' window. 
This also flushes out stored/errant clicks that occurred while Synaptic or the 'View and Upgrade' window were open.